@@ -2,30 +2,31 @@
   "name": "sharp-image",
   "version": "0.0.1",
   "private": true,
+  "type": "module",
   "scripts": {
-    "start": "node --use_strict --harmony ."
+    "start": "node .",
+    "lint": "eslint ."
   },
   "main": "app.js",
   "env": {
-    "NODE_ENV": "production"
+    "NODE_ENV": "development"
   },
   "dependencies": {
-    "async": "^2.1.4",
-    "cron": "^1.2.1",
-    "file-type": "^4.1.0",
-    "lodash": "^4.17.4",
-    "mongodb": "^2.2.22",
-    "sharp": "^0.17.2",
-    "ua-parser-js": "^0.7.12"
+    "cron": "^2.1.0",
+    "file-type": "^18.0.0",
+    "mongodb": "^4.12.1",
+    "p-limit": "^4.0.0",
+    "sharp": "^0.31.2",
+    "ua-parser-js": "^1.0.32"
   },
   "description": "Image cdn server",
   "author": "Akhilesh Jain",
   "license": "ISC",
   "devDependencies": {
-    "eslint": "^3.15.0",
-    "eslint-config-airbnb": "^14.1.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "^6.9.0"
+    "eslint": "^8.28.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-react": "^7.31.11"
   }
 }
